Validação dos teste, e verificação nos metodos para evitar falhas

+++ backend/controllers/base_controller.py
@@ -1,11 +1,18 @@
 from backend.models.base_model import BaseModel
+from backend.dao.base_dao import BaseDao
 class BaseController:
     def __init__(self, dao):
-        self.__dao = dao
+        if isinstance(dao, BaseDao):
+            self.__dao = dao
+        else:
+            raise ValueError('dao must receive a valid dao')
+        
 
     def save(self, model: BaseModel) -> None:
-        self.__dao.save(model)
-
+        if isinstance(model, BaseModel):
+            self.__dao.save(model)
+        else:
+            raise ValueError('the save method must receive object of a Class')
     def read_by_id(self, id: int) -> BaseModel:
         objetc_by_id = self.__dao.read_by_id(id)
         return objetc_by_id
@@ -15,4 +22,7 @@ class BaseController:
         return list_all
 
     def delete(self, model: BaseModel) -> None:
-        self.__dao.delete(model)
+        if isinstance(model, BaseModel):
+            self.__dao.delete(model)
+        else:
+            raise ValueError("The save method must receive object of a Class")
